+++ operator/app/src/test/java/org/keycloak/operator/testsuite/integration/KeycloakServicesTest.java
@@ -1,4 +1,21 @@
-package org.keycloak.operator;
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.keycloak.operator.testsuite.integration;
 
 import io.fabric8.kubernetes.api.model.ServiceSpecBuilder;
 import io.quarkus.logging.Log;
@@ -7,14 +24,14 @@ import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
 import org.keycloak.operator.controllers.KeycloakDiscoveryService;
 import org.keycloak.operator.controllers.KeycloakService;
-import org.keycloak.operator.utils.K8sUtils;
+import org.keycloak.operator.testsuite.utils.K8sUtils;
 
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @QuarkusTest
-public class KeycloakServicesE2EIT extends ClusterOperatorTest {
+public class KeycloakServicesTest extends BaseOperatorTest {
     @Test
     public void testMainServiceDurability() {
         var kc = K8sUtils.getDefaultKeycloakDeployment();
