+++ operator/app/src/test/java/org/keycloak/operator/testsuite/integration/PodTemplateTest.java
@@ -1,4 +1,21 @@
-package org.keycloak.operator;
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.keycloak.operator.testsuite.integration;
 
 import io.fabric8.kubernetes.api.model.PodTemplateSpecBuilder;
 import io.fabric8.kubernetes.client.dsl.Resource;
@@ -7,7 +24,7 @@ import io.quarkus.logging.Log;
 import io.quarkus.test.junit.QuarkusTest;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
-import org.keycloak.operator.utils.CRAssert;
+import org.keycloak.operator.testsuite.utils.CRAssert;
 import org.keycloak.operator.crds.v2alpha1.deployment.Keycloak;
 
 import static java.util.concurrent.TimeUnit.MINUTES;
@@ -15,7 +32,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.keycloak.operator.crds.v2alpha1.deployment.KeycloakStatusCondition.HAS_ERRORS;
 
 @QuarkusTest
-public class PodTemplateE2EIT extends ClusterOperatorTest {
+public class PodTemplateTest extends BaseOperatorTest {
 
     private Keycloak getEmptyPodTemplateKeycloak() {
         return Serialization.unmarshal(getClass().getResourceAsStream("/empty-podtemplate-keycloak.yml"), Keycloak.class);
